% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs.R
\name{concat.ms}
\alias{concat.ms}
\title{Concat MS}
\usage{
concat.ms(...)
}
\arguments{
\item{...}{MS objects to concatenate}
}
\value{
MS object containing all input objects'
}
\description{
Concatinate two or more MS objects
}

