% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs.R
\name{length.ms}
\alias{length.ms}
\title{Length of MS object}
\usage{
\method{length}{ms}(x)
}
\arguments{
\item{x}{MS object}
}
\value{
Integer containing the number of sequences the MS object contains
}
\description{
Return number of sequences contained in MS object
}

