\name{lengths.ms}
\alias{lengths.ms}
\title{MS sequence lengths}
\usage{lengths.ms(x)
}
\description{Returns the length of each sequence in an MS object.}
\value{integer vector giving the length of each sequence in the MS object}
\keyword{ms}
\author{Nick Peterson}
\arguments{\item{x}{an MS object}
}

