\name{names.ms}
\alias{names.ms}
\title{MS Sequence Names}
\usage{names.ms(x)
}
\description{Returns a list of sequence names contained in an an MS object.}
\value{a character vector giving the names of the sequences, or
NULL if they are not defined}
\keyword{ms}
\author{Nick Peterson}
\arguments{\item{x}{an MS object}
}

