\name{names.ms}
\alias{names.ms}
\title{MS Sequence Names}
\usage{
  \method{names}{ms} (x)
}
\arguments{
  \item{x}{an MS object}
}
\value{
  a character vector giving the names of the sequences, or
  NULL if they are not defined
}
\description{
  Returns a list of sequence names contained in an an MS
  object.
}
\author{
  Nick Peterson
}
\keyword{ms}

