\name{print.ms}
\alias{print.ms}
\title{Printing MS objects}
\usage{print.ms(x, ..., print.all=((length(x) < 15) && (sum(lengths.ms(x)) < 500)))
}
\description{Prints an MS (multiple sequence) object.}
\keyword{ms}
\author{Nick Peterson}
\arguments{\item{x}{an object of class ms}
\item{...}{additional arguments sent to \code{print}}
\item{print.all}{whether to suppress printing of the bases, names, offsets}
}

