\name{read.mm}
\alias{read.mm}
\title{Read Markov Model from file}
\usage{read.mm(filename)
}
\description{Read in a Markov model from file created by MEME's fasta-get-markov tool}
\value{Markov Model (list of Markov Matrices, one for each order)}
\arguments{\item{filename}{Full path to output file from fasta-get-markov}
}

