% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs.R
\name{read.mm}
\alias{read.mm}
\title{Read Markov Model from file}
\usage{
read.mm(filename)
}
\arguments{
\item{filename}{Full path to output file from fasta-get-markov}
}
\value{
Markov Model (list of Markov Matrices, one for each order)
}
\description{
Read in a Markov model from file created by MEME's fasta-get-markov tool
}

