% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs.R
\name{sequences.ms}
\alias{sequences.ms}
\title{Get sequences}
\usage{
sequences.ms(ms)
}
\arguments{
\item{ms}{An MS object}
}
\value{
character vector containing the sequences in the MS object
}
\description{
Returns list of sequences in an MS object
}

