\name{write.ms}
\alias{write.ms}
\title{Writing MS Object to FASTA file}
\usage{
write.ms(x, file=NULL)
}
\description{Writes a multiple sequence (MS) object to a FASTA file.}
\keyword{ms}
\keyword{FASTA}
\author{Nick Peterson}
\arguments{\item{x}{an object of class ms}
\item{file}{File to write (will be overwritten).  If NULL, output
goes to terminal.}
}

