% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tfbs.R
\name{write.ms}
\alias{write.ms}
\title{Writing MS Object to FASTA file}
\usage{
write.ms(x, file=NULL)
}
\arguments{
\item{x}{an object of class ms}

\item{file}{File to write (will be overwritten).  If NULL, output
goes to terminal.}
}
\description{
Writes a multiple sequence (MS) object to a FASTA file.
}
\author{
Nick Peterson
}
\keyword{FASTA}
\keyword{ms}

