% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs.R
\name{lengths.ms}
\alias{lengths.ms}
\title{MS sequence lengths}
\usage{
lengths.ms(x)
}
\arguments{
\item{x}{an MS object}
}
\value{
integer vector giving the length of each sequence in the MS object
}
\description{
Returns the length of each sequence in an MS object.
}
\author{
Nick Peterson
}
\keyword{ms}

